# book-network/src/main/java/com/mlk/book/feedback/FeedbackResponse.java
package com.mlk.book.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}